+++ js/util.js
@@ -1,0 +1,6 @@
+function getRandomNumber(a, b) {
+  return Math.floor(Math.random() * (b - a)) + a;
+}
+const stringLength = (str, maxLenght) => str.length <= maxLenght;
+
+export {getRandomNumber, stringLength};
